Add files via upload
Although this was a pretty easy platform to use, I still found it frustrating how clunky the characters could be. Of course, I understand that in order to create a less clunky game, we need to be able to use less drag and drop functions and more hard core coding.